Create a File for key settings
Build a common file that initializes the root window and the StringVars so keys_settings and main can import these without calling each other
